Split to multiples classes. Split parsing command line to new method. Add structs to save command line parameters.
