+++ INPTPZ1/INPTPZ1.csproj
@@ -43,8 +43,11 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Cplx.cs" />
+    <Compile Include="Poly.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Properties\BitmapParameters.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
